requirements.txt
Django
djangorestframework
Pillow

